+++ profiles/elvui.lua
@@ -2852,6 +2852,7 @@ local function SetupTrenchyUI()
     E.db.TrenchyUI.addons.skinPlatynator = true
     E.db.TrenchyUI.addons.skinWarpDeplete = true
     E.db.TrenchyUI.addons.skinWorldQuestTab = true
+    E.db.TrenchyUI.addons.skinPremadeGroupsFilter = true
     E.db.TrenchyUI.pixelGlow.enabled = true
     E.db.TrenchyUI.pixelGlow.thickness = 3
     E.db.TrenchyUI.fakePower.soulFragments = true